Add files via upload
This project is a simple web application that allows users to input tasks and display them in a list format on the screen. Users can add tasks by entering them in an input box and clicking a button. The added tasks will be dynamically displayed on the screen. Additionally, there is a delete button associated with each task that allows users to remove tasks from the list.